requirements.txt
pytest==7.4.3
pytest-asyncio==0.23.2
python-telegram-bot==20.7 